refactor(aurora-portal): remake sec-groups and fl-ip route structure
