+++ apps/aurora-portal/src/client/routes/_auth/accounts/$accountId/projects/$projectId/network/securitygroups/$securityGroupId/index.tsx
@@ -5,7 +5,7 @@ import { Trans, useLingui } from "@lingui/react/macro"
 import { useMemo } from "react"
 import { getServiceIndex } from "@/server/Authentication/helpers"
 import { SecurityGroupDetailsView } from "./-components/SecurityGroupDetailsView"
-import { EditSecurityGroupModal } from "../../-components/SecurityGroups/-components/-modals/EditSecurityGroupModal"
+import { EditSecurityGroupModal } from "../-components/-modals/EditSecurityGroupModal"
 import { useSecurityGroupDetails } from "./-hooks/useSecurityGroupDetails"
 import { useListWithFiltering } from "@/client/utils/useListWithFiltering"
 import { trpcReact } from "@/client/trpcClient"
