+++ apps/aurora-portal/src/client/routes/_auth/accounts/$accountId/projects/$projectId/network/securitygroups/-components/-modals/CreateSecurityGroupModal.test.tsx
@@ -8,7 +8,6 @@ import { CreateSecurityGroupModal } from "./CreateSecurityGroupModal"
 import { CreateSecurityGroupInput } from "@/server/Network/types/securityGroup"
 
 // ─── Render helper ────────────────────────────────────────────────────────────
-
 const renderModal = ({
   isOpen = true,
   onClose = vi.fn(),
@@ -29,7 +28,6 @@ const renderModal = ({
   )
 
 // ─── Tests ────────────────────────────────────────────────────────────────────
-
 describe("CreateSecurityGroupModal", () => {
   beforeEach(async () => {
     vi.clearAllMocks()
