+++ apps/aurora-portal/src/client/routes/_auth/accounts/$accountId/projects/$projectId/network/securitygroups/-components/-modals/CreateSecurityGroupModal.tsx
