+++ apps/aurora-portal/src/server/Compute/helpers/imageHelpers.test.ts
@@ -1072,7 +1072,7 @@ describe("imageHelpers", () => {
         {
           code: "invalid_type",
           expected: "string",
-          received: "number",
+          input: 123,
           path: ["name"],
           message: "Expected string, received number",
         },
